Refactor component names and delete unused components
