+++ src/components/atoms/vCourseDeleteAvatar.vue
