+++ src/components/molecules/vCourseGroupAvatar.unit.ts
