+++ src/components/molecules/vCourseGroupAvatar.vue
